Back

+++ back/app/Http/Controllers/CardController.php
@@ -2,8 +2,15 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Card;
-use App\Models\MealPrice;
+use App\Http\Requests\{
+    CardStoreRequest,
+    CardUpdateRequest
+};
+use App\Models\{
+    AccomodationPayment,
+    Card,
+    MealPrice
+};
 use Illuminate\Http\Request;
 
 class CardController extends Controller
@@ -11,16 +18,18 @@ class CardController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  App\Http\Request\CardStoreRequest $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CardStoreRequest $request)
     {
+        $this->authorize('create', Card::class);
+
         return response()->json(Card::store([
-            "creation_date"=>$request->input('creation_date'),
-            "expiration_date"=>$request->input('expiration_date'),
-            "card_type_id"=>$request->input('card_type_id'),
-            "user_id"=>$request->input('user_id')
+            "creation_date" => $request->input('creation_date'),
+            "expiration_date" => $request->input('expiration_date'),
+            "card_type_id" => $request->input('card_type_id'),
+            "user_id" => $request->input('user_id')
         ]));
     }
 
@@ -32,39 +41,51 @@ class CardController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Card $card)
     {
-        return response()->json(Card::find($id));
+        $this->authorize('view', $card);
+
+        return response()->json($card);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  App\Http\Requests\CardUpdateRequest  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(CardUpdateRequest $request, $id)
     {
+        $this->authorize('update', Card::class);
+
         $card=Card::find($id);
+        
         if(empty($request->input('money'))
             &&empty($request->input('breakfast'))
             &&empty($request->input('lunch'))
             &&empty($request->input('dinner')))
-        return response()->json($card->update($request->all()));
+            return response()->json($card->update($request->all()));
     }
 
-    public function addMeals(Request $request,$id)
+    public function addMeals(Request $request, Card $card)
     {
-        $card=Card::find($id);
-        $mealPrices=MealPrice::select('price','meal_id')->where('card_type_id','=',$card->card_type_id);
-        $money=$request->input('breakfast')*$mealPrices[0]->price+$request->input('lunch')*$mealPrices[1]->price+$request->input('dinner')*$mealPrices[2]->price;
-        if($card->money>=$money){
-            $card->breakfast+=$request->input('breakfast');
-            $card->lunch+=$request->input('lunch');
-            $card->dinner+=$request->input('dinner');
-            $card->money-=$money;
+        $this->authorize('addMeal', $card);
+
+        $mealPrices=MealPrice::select('price', 'meal_id')->where('card_type_id', '=', $card->card_type_id);
+        
+        $money=$request->input('breakfast') * $mealPrices[0]->price
+                +$request->input('lunch') * $mealPrices[1]->price
+                +$request->input('dinner') * $mealPrices[2]->price;
+
+        if($card->money >= $money) {
+            $card->breakfast += $request->input('breakfast');
+            $card->lunch += $request->input('lunch');
+            $card->dinner += $request->input('dinner');
+            $card->money -= $money;
         }
+        
+        return response()->json($card);
     }
 
     /**
@@ -75,6 +96,8 @@ class CardController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Card::class);
+        
         return response()->json(Card::destroy($id));
     }
 }
