+++ back/app/Http/Controllers/TicketController.php
@@ -2,6 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\{
+    TicketStoreRequest,
+    TicketUpdateRequest
+};
 use App\Models\Ticket;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
@@ -15,7 +19,13 @@ class TicketController extends Controller
      */
     public function index()
     {
-        return response()->json([Ticket::select('time')->orderByDesc('time')->get()]);
+        $this->authorize('viewAny', Ticket::class);
+
+        return response()->json([
+            Ticket::select('time')
+                ->orderByDesc('time')
+                ->get()
+        ]);
     }
 
     /**
@@ -24,29 +34,35 @@ class TicketController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function ticketResponse(Request $request,$id)
+    public function ticketResponse(Request $request, $id)
     {
+        $this->authorize('response', Ticket::class);
+
         return response()->json(Ticket::create([
-            "content"=>$request->input('content'),
-            "time"=>$request->input('time'),
-            "user_id"=>$request->input('user_id'),
-            "parent_id"=>$id
+            "content" => $request->input('content'),
+            "time" => $request->input('time'),
+            "user_id" => $request->input('user_id'),
+            "parent_id" => $id
         ]));
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  App\Http\Requests\TicketStoreRequest $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(TicketStoreRequest $request)
     {
-        return response()->json(Ticket::create([
-            "content"=>$request->input('content'),
-            "time"=>$request->input('time'),
-            "user_id"=>$request->input('user_id')
-        ]));
+        $this->authorize('create', Ticket::class);
+
+        return response()->json(
+            Ticket::create([
+            "content" => $request->input('content'),
+            "time" => $request->input('time'),
+            "user_id" => $request->input('user_id')
+            ])
+        );
     }
 
     /**
@@ -55,34 +71,44 @@ class TicketController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Ticket $ticket)
     {
-        $ticket=DB::table('tickets')
-        ->join('users','tickets.user_id','=','users.id')
-        ->select('tickets.content','tickets.time','users.name','users.lastname')
-        ->where('tickets.id','=',$id)
-        ->limit(1)
-        ->get();
+        $this->authorize('view', $ticket);
+
+        $tckt=DB::table('tickets')
+            ->join('users', 'tickets.user_id', '=', 'users.id')
+            ->select('tickets.content', 'tickets.time', 'users.name', 'users.lastname')
+            ->where('tickets.id', '=', $ticket->id)
+            ->limit(1)
+            ->get();
+
         $response=DB::table('tickets')
-        ->join('users','tickets.user_id','=','users.id')
-        ->select('tickets.content','tickets.time','users.name','users.lastname')
-        ->where('tickets.parent_id','=',$id)
-        ->limit(1)
-        ->get();
-        return response()->json([$ticket,$response]);
+            ->join('users', 'tickets.user_id', '=', 'users.id')
+            ->select('tickets.content', 'tickets.time', 'users.name', 'users.lastname')
+            ->where('tickets.parent_id', '=', $ticket->id)
+            ->limit(1)
+            ->get();
+
+        return response()->json([
+            $tckt,
+            $response
+        ]);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  App\Http\Requests\TicketUpdateRequest  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(TicketUpdateRequest $request, Ticket $ticket)
     {
-        $ticket=Ticket::find($id);
-        return response()->json($ticket->update($request->all()));
+        $this->authorize('update', $ticket);
+        
+        return response()->json(
+            $ticket->update($request->all())
+        );
     }
 
     /**
@@ -93,6 +119,8 @@ class TicketController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete',Ticket::class);
+        
         return response()->json(Ticket::destroy($id));
     }
 }
