+++ back/app/Http/Controllers/UserController.php
@@ -2,6 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\{
+    UserStoreRequest,
+    UserUpdateRequest
+};
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -11,19 +15,23 @@ class UserController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  App\Http\Requests\UserStoreRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(UserStoreRequest $request)
     {
-        return response()->json(User::create([
-            'name'=>$request->input('name'),
-            'lastname'=>$request->input('lastname'),
-            'email'=>$request->input('email'),
-            'password'=>bcrypt($request->input('password')),
-            'year_of_birth'=>$request->input('year_of_birth'),
-            'role_id'=>$request->input('role_id')
-        ]));
+        $this->authorize('create', User::class);
+
+        return response()->json(
+            User::create([
+                'name' => $request->input('name'),
+                'lastname' => $request->input('lastname'),
+                'email' => $request->input('email'),
+                'password' => bcrypt($request->input('password')),
+                'year_of_birth' => $request->input('year_of_birth'),
+                'role_id' => $request->input('role_id')
+            ])
+        );
     }
 
     /**
@@ -32,9 +40,11 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(User $user)
     {
-        return response()->json(User::find($id));
+        $this->authorize('view', $user);
+
+        return response()->json($user);
     }
 
     public function showLoggedUser()
@@ -45,25 +55,29 @@ class UserController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  App\Http\Requests\UserUpdateRequest $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(UserUpdateRequest $request, $id)
     {
+        $this->authorize('update', User::class);
         $user=User::find($id);
         if(empty($request->input('password')))    
             $user->update($request->all());
-        else{
+        else {
             $user->update($request->except('password'));
         }
         return response()->json($user);
     }
 
-    public function changePassword(Request $request,$id)
+    public function changePassword(Request $request, $id)
     {
         $user=User::find($id);
-        return response()->json($user->update(["password"=>bcrypt($request->input('password'))]));
+        return response()->json(
+            $user->update([
+                "password"=>bcrypt($request->input('password'))
+            ]));
     }
 
     /**
@@ -74,6 +88,8 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('destroy', User::class);
+
         return response()->json(User::destroy($id));
     }
 }
